Made a number of changes aimed at reconciling cell comments and form controls in keep_vba mode.
In particular, if a worksheet has a legacyDrawing link to a vml file then
the vml parts of any new comments are merged into the existing vml
file rather than being written out in a new vml file.

There are also a few changes to names of variables and links to reflect the
fact that a worksheet has a single vml file containing descriptions of
controls and comments.

--HG--
branch : 2.3